chore: move the settings component to be exported directly in the index file
