--- src/components/tx/confirmation-views/SettingsChange/index.ts
@@ -1,3 +1,0 @@
-import SettingsChange from './SettingsChange'
-
-export default SettingsChange
