+++ src/components/tx/confirmation-views/SettingsChange/index.tsx
